enhancement (search module): create header calendar component
