+++ projects/knora/search/src/lib/extended-search/select-property/specify-property-value/date-value/header-calendar/header-calendar.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { HeaderComponent } from './header-calendar.component';
+
+describe('HeaderComponent', () => {
+  let component: HeaderComponent<any>;
+  let fixture: ComponentFixture<HeaderComponent<any>>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [HeaderComponent]
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(HeaderComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
